Finalizing of methods name
