+++ app/views/decks/_decks.html.erb
@@ -3,7 +3,7 @@
   <div class="col-md-3">
     <div class="panel panel-primary">
       <div class="panel-heading">
-        <span class="panel-title"><%= link_to deck.name, deck %></span>
+        <span class="panel-title"><%= link_to deck.name, deck %> (<%= deck.cards.count %>)</span>
         <span class="pull-right">#<%= deck.id %></span>
         <%= link_to(deck, method: :delete, data: { confirm: t('menu.confirm') }, title: t('menu.delete')) do %>
           <span class="glyphicon glyphicon-remove white close" aria-hidden="true"></span>
