--- db/structure.sql
@@ -1,334 +1,0 @@
---
--- PostgreSQL database dump
---
-
-SET statement_timeout = 0;
-SET lock_timeout = 0;
-SET client_encoding = 'UTF8';
-SET standard_conforming_strings = on;
-SET check_function_bodies = false;
-SET client_min_messages = warning;
-
---
--- Name: plpgsql; Type: EXTENSION; Schema: -; Owner: -
---
-
-CREATE EXTENSION IF NOT EXISTS plpgsql WITH SCHEMA pg_catalog;
-
-
---
--- Name: EXTENSION plpgsql; Type: COMMENT; Schema: -; Owner: -
---
-
-COMMENT ON EXTENSION plpgsql IS 'PL/pgSQL procedural language';
-
-
-SET search_path = public, pg_catalog;
-
-SET default_tablespace = '';
-
-SET default_with_oids = false;
-
---
--- Name: authentications; Type: TABLE; Schema: public; Owner: -; Tablespace: 
---
-
-CREATE TABLE authentications (
-    id integer NOT NULL,
-    user_id integer NOT NULL,
-    provider character varying NOT NULL,
-    uid character varying NOT NULL,
-    created_at timestamp without time zone,
-    updated_at timestamp without time zone
-);
-
-
---
--- Name: authentications_id_seq; Type: SEQUENCE; Schema: public; Owner: -
---
-
-CREATE SEQUENCE authentications_id_seq
-    START WITH 1
-    INCREMENT BY 1
-    NO MINVALUE
-    NO MAXVALUE
-    CACHE 1;
-
-
---
--- Name: authentications_id_seq; Type: SEQUENCE OWNED BY; Schema: public; Owner: -
---
-
-ALTER SEQUENCE authentications_id_seq OWNED BY authentications.id;
-
-
---
--- Name: cards; Type: TABLE; Schema: public; Owner: -; Tablespace: 
---
-
-CREATE TABLE cards (
-    id integer NOT NULL,
-    original_text character varying,
-    translated_text character varying,
-    review_date timestamp without time zone,
-    created_at timestamp without time zone NOT NULL,
-    updated_at timestamp without time zone NOT NULL,
-    picture_file_name character varying,
-    picture_content_type character varying,
-    picture_file_size integer,
-    picture_updated_at timestamp without time zone,
-    deck_id integer
-);
-
-
---
--- Name: cards_id_seq; Type: SEQUENCE; Schema: public; Owner: -
---
-
-CREATE SEQUENCE cards_id_seq
-    START WITH 1
-    INCREMENT BY 1
-    NO MINVALUE
-    NO MAXVALUE
-    CACHE 1;
-
-
---
--- Name: cards_id_seq; Type: SEQUENCE OWNED BY; Schema: public; Owner: -
---
-
-ALTER SEQUENCE cards_id_seq OWNED BY cards.id;
-
-
---
--- Name: decks; Type: TABLE; Schema: public; Owner: -; Tablespace: 
---
-
-CREATE TABLE decks (
-    id integer NOT NULL,
-    name character varying,
-    created_at timestamp without time zone NOT NULL,
-    updated_at timestamp without time zone NOT NULL,
-    user_id integer
-);
-
-
---
--- Name: decks_id_seq; Type: SEQUENCE; Schema: public; Owner: -
---
-
-CREATE SEQUENCE decks_id_seq
-    START WITH 1
-    INCREMENT BY 1
-    NO MINVALUE
-    NO MAXVALUE
-    CACHE 1;
-
-
---
--- Name: decks_id_seq; Type: SEQUENCE OWNED BY; Schema: public; Owner: -
---
-
-ALTER SEQUENCE decks_id_seq OWNED BY decks.id;
-
-
---
--- Name: schema_migrations; Type: TABLE; Schema: public; Owner: -; Tablespace: 
---
-
-CREATE TABLE schema_migrations (
-    version character varying NOT NULL
-);
-
-
---
--- Name: stable; Type: TABLE; Schema: public; Owner: -; Tablespace: 
---
-
-CREATE TABLE stable (
-    yo boolean
-);
-
-
---
--- Name: users; Type: TABLE; Schema: public; Owner: -; Tablespace: 
---
-
-CREATE TABLE users (
-    id integer NOT NULL,
-    email character varying,
-    crypted_password character varying,
-    salt character varying,
-    created_at timestamp without time zone,
-    updated_at timestamp without time zone,
-    current_deck_id integer
-);
-
-
---
--- Name: users_id_seq; Type: SEQUENCE; Schema: public; Owner: -
---
-
-CREATE SEQUENCE users_id_seq
-    START WITH 1
-    INCREMENT BY 1
-    NO MINVALUE
-    NO MAXVALUE
-    CACHE 1;
-
-
---
--- Name: users_id_seq; Type: SEQUENCE OWNED BY; Schema: public; Owner: -
---
-
-ALTER SEQUENCE users_id_seq OWNED BY users.id;
-
-
---
--- Name: id; Type: DEFAULT; Schema: public; Owner: -
---
-
-ALTER TABLE ONLY authentications ALTER COLUMN id SET DEFAULT nextval('authentications_id_seq'::regclass);
-
-
---
--- Name: id; Type: DEFAULT; Schema: public; Owner: -
---
-
-ALTER TABLE ONLY cards ALTER COLUMN id SET DEFAULT nextval('cards_id_seq'::regclass);
-
-
---
--- Name: id; Type: DEFAULT; Schema: public; Owner: -
---
-
-ALTER TABLE ONLY decks ALTER COLUMN id SET DEFAULT nextval('decks_id_seq'::regclass);
-
-
---
--- Name: id; Type: DEFAULT; Schema: public; Owner: -
---
-
-ALTER TABLE ONLY users ALTER COLUMN id SET DEFAULT nextval('users_id_seq'::regclass);
-
-
---
--- Name: authentications_pkey; Type: CONSTRAINT; Schema: public; Owner: -; Tablespace: 
---
-
-ALTER TABLE ONLY authentications
-    ADD CONSTRAINT authentications_pkey PRIMARY KEY (id);
-
-
---
--- Name: cards_pkey; Type: CONSTRAINT; Schema: public; Owner: -; Tablespace: 
---
-
-ALTER TABLE ONLY cards
-    ADD CONSTRAINT cards_pkey PRIMARY KEY (id);
-
-
---
--- Name: decks_pkey; Type: CONSTRAINT; Schema: public; Owner: -; Tablespace: 
---
-
-ALTER TABLE ONLY decks
-    ADD CONSTRAINT decks_pkey PRIMARY KEY (id);
-
-
---
--- Name: users_pkey; Type: CONSTRAINT; Schema: public; Owner: -; Tablespace: 
---
-
-ALTER TABLE ONLY users
-    ADD CONSTRAINT users_pkey PRIMARY KEY (id);
-
-
---
--- Name: index_authentications_on_provider_and_uid; Type: INDEX; Schema: public; Owner: -; Tablespace: 
---
-
-CREATE INDEX index_authentications_on_provider_and_uid ON authentications USING btree (provider, uid);
-
-
---
--- Name: index_cards_on_deck_id; Type: INDEX; Schema: public; Owner: -; Tablespace: 
---
-
-CREATE INDEX index_cards_on_deck_id ON cards USING btree (deck_id);
-
-
---
--- Name: index_decks_on_user_id; Type: INDEX; Schema: public; Owner: -; Tablespace: 
---
-
-CREATE INDEX index_decks_on_user_id ON decks USING btree (user_id);
-
-
---
--- Name: index_users_on_current_deck_id; Type: INDEX; Schema: public; Owner: -; Tablespace: 
---
-
-CREATE INDEX index_users_on_current_deck_id ON users USING btree (current_deck_id);
-
-
---
--- Name: index_users_on_email; Type: INDEX; Schema: public; Owner: -; Tablespace: 
---
-
-CREATE UNIQUE INDEX index_users_on_email ON users USING btree (email);
-
-
---
--- Name: unique_schema_migrations; Type: INDEX; Schema: public; Owner: -; Tablespace: 
---
-
-CREATE UNIQUE INDEX unique_schema_migrations ON schema_migrations USING btree (version);
-
-
---
--- Name: fk_rails_5d31349cbe; Type: FK CONSTRAINT; Schema: public; Owner: -
---
-
-ALTER TABLE ONLY decks
-    ADD CONSTRAINT fk_rails_5d31349cbe FOREIGN KEY (user_id) REFERENCES users(id);
-
-
---
--- Name: fk_rails_6c4effce17; Type: FK CONSTRAINT; Schema: public; Owner: -
---
-
-ALTER TABLE ONLY cards
-    ADD CONSTRAINT fk_rails_6c4effce17 FOREIGN KEY (deck_id) REFERENCES decks(id);
-
-
---
--- PostgreSQL database dump complete
---
-
-SET search_path TO "$user",public;
-
-INSERT INTO schema_migrations (version) VALUES ('20150819132046');
-
-INSERT INTO schema_migrations (version) VALUES ('20150903190400');
-
-INSERT INTO schema_migrations (version) VALUES ('20150903223824');
-
-INSERT INTO schema_migrations (version) VALUES ('20150907101105');
-
-INSERT INTO schema_migrations (version) VALUES ('20150907101704');
-
-INSERT INTO schema_migrations (version) VALUES ('20150907104604');
-
-INSERT INTO schema_migrations (version) VALUES ('20150907202310');
-
-INSERT INTO schema_migrations (version) VALUES ('20150914115903');
-
-INSERT INTO schema_migrations (version) VALUES ('20151003120648');
-
-INSERT INTO schema_migrations (version) VALUES ('20151003123622');
-
-INSERT INTO schema_migrations (version) VALUES ('20151003123801');
-
-INSERT INTO schema_migrations (version) VALUES ('20151006193601');
-
